feat: logique and style ok
